tweaks

+++ Assets/Shaders/MyShader_Code.shader
@@ -53,7 +53,6 @@ Shader "Custom/MyShader_Code"
                 {
                     const int label_index = channel * 8 + bit;
                     const int mask = 1 << bit;
-                    // const int color32 = round(vColor[channel] * 255);
                     const float f = vColor[channel];
                     const int color32 = f >= 1.0 ? 255 : f <= 0.0 ? 0 : (int)floor(f * 256.0);
                     const int result = mask & color32;
@@ -64,7 +63,8 @@ Shader "Custom/MyShader_Code"
                         const float4 label_color = _ColorArray[label_index];
                         if (label_color.a >= 0.5) // Check if the label is visible
                         {
-                        final_color *= label_color;
+                            final_color *= label_color;
+                            // final_color = lerp(final_color, label_color, label_color.a);
                         }
                     }
                 }
@@ -82,7 +82,7 @@ Shader "Custom/MyShader_Code"
             // Combined labels color
             float4 labelColor = CombineLabels(IN.color);
 
-            float4 c = mainTexColor * labelColor;
+            fixed4 c = mainTexColor * labelColor;
 
             o.Albedo = c.rgb;
             // Metallic and smoothness come from slider variables
